requirements.txt
fastapi
uvicorn
websockets
nest_asyncio
python-dotenv
groq
PyMuPDF
langchain
streamlit
sentence-transformers
faiss-cpu
python-multipart